+++ application/views/_partial/collection/category-layout-3.php
@@ -1,6 +1,6 @@
 <?php defined('SYSTEM_INIT') or die('Invalid Usage.'); ?>
 <?php if (isset($collection['categories']) && count($collection['categories'])) { ?>
-    <section class="section">
+    <section class="section" data-section="section">
         <div class="container">
             <div class="section-head section-head-center">
                 <?php echo ($collection['collection_name'] != '') ? ' <div class="section-heading"><h2>' . $collection['collection_name'] . '</h2></div>' : ''; ?>
@@ -35,17 +35,19 @@
                         <div class="category-body">
                             <ul class="category-list">
                                 <li class="category-list-item category-list-head">
-                                    <a href="<?php echo UrlHelper::generateUrl('Category', 'View', array($category['prodcat_id'])); ?>">
+                                    <a
+                                        href="<?php echo UrlHelper::generateUrl('Category', 'View', array($category['prodcat_id'])); ?>">
                                         <?php echo $category['prodcat_name']; ?>
                                     </a>
                                 </li>
                                 <?php $i = 1;
                                 foreach ($category['subCategories'] as $subCat) { ?>
                                     <li class="category-list-item">
-                                        <a href="<?php echo UrlHelper::generateUrl('Category', 'View', array($subCat['prodcat_id'])); ?>">
+                                        <a
+                                            href="<?php echo UrlHelper::generateUrl('Category', 'View', array($subCat['prodcat_id'])); ?>">
                                             <?php echo $subCat['prodcat_name']; ?></a>
                                     </li>
-                                <?php $i++;
+                                    <?php $i++;
                                     if ($i > 5) {
                                         break;
                                     }
@@ -58,7 +60,8 @@
             </div>
             <?php if (count($collection['categories']) > Collections::LIMIT_CATEGORY_LAYOUT3) { ?>
                 <div class="section-foot">
-                    <a href="<?php echo UrlHelper::generateUrl('Collections', 'View', array($collection['collection_id'])); ?>" class="link-underline">
+                    <a href="<?php echo UrlHelper::generateUrl('Collections', 'View', array($collection['collection_id'])); ?>"
+                        class="link-underline">
                         <?php echo Labels::getLabel('LBL_VIEW_ALL', $siteLangId); ?>
                     </a>
                 </div>