+++ application/views/restore-system/top-header.php
@@ -2,8 +2,11 @@
     <svg class="svg" width="24" height="24">
         <use xlink:href="<?php echo CONF_WEBROOT_FRONTEND; ?>images/retina/sprite-restore.svg#gear">
         </use>
-    </svg>
+    </svg>    
 </button>
+<div class="fixed-right-btn">
+    <button type="button" class="sticker-Btn popup-btn" onClick="demoRequestForm()"><span class="arrow-btn"><i class="arrow-anim"></i></span>BOOK A DEMO</>
+</div>
 <div class="offcanvas offcanvas-end offcanvas-restore" tabindex="-1" id="offcanvasRestore">
     <div class="offcanvas-body">
         <div class="demo">