+++ application/views/testimonials/index.php
@@ -1,7 +1,7 @@
 <?php defined('SYSTEM_INIT') or die('Invalid Usage.'); ?>
 <div id="body" class="body">
     <?php $this->includeTemplate('_partial/page-head-section.php', ['headLabel' => Labels::getLabel('LBL_TESTIMONIALS'), 'includeBreadcrumb' => true]); ?>
-    <section class="section">
+    <section class="section" data-section="section">
         <div class="container">
             <div class="cms">
                 <div class="list__all" id='listing'></div>