+++ CHANGELOG.md
@@ -17,6 +17,11 @@
 ## Enhancements:
     - Load all categories on ajax call instead of page load.
     - Performance optimization to handle product ratings.
+    - Geo location popup at the header section.
+    - Language and currency selection via popup.
+    - Displayed review images at the top of reviews list.
+    - Performance optimization ar admin end, product search, filters and collections list.
+    
 
 ## Bug Fixes:
     - Bug-074802 - Omise Payment gateway deprecated issue.
@@ -85,5 +90,7 @@
     - Bug-085177 - 'Final Acceptance button does not appear to seller when buyer accepts the offer of seller
     - Bug-085157 - 'Request for Quote' button is displayed on the bottom of the home page also
 
+    - Bug-085143 - 404 on the product details page when the new user creates a product.
+    - Bug-085068 - cancellation is not working properly if stripe connect is on.
 ## Hit URL:
     - admin/patch-update/updateProductRating